+++ src/view/components/LogoMinified.js
@@ -3,20 +3,20 @@ import { useHistory } from "react-router-dom";
 
 import "../styles/logo.css";
 
-import logo from "../images/logo-styled.svg";
+import logoMinified from "../images/logo-without-background.svg";
 import { siteName } from "../../model/adminAssets.json";
 
 export default function Logo() {
   const history = useHistory();
 
-  function goToHomePage(e) {
+  const goToHomePage = (e) => {
     e.preventDefault();
     history.push("/");
-  }
+  };
 
   return (
     <img
-      src={logo}
+      src={logoMinified}
       id="logo"
       className="img-fluid"
       alt={siteName}