+++ src/view/components/MainMenuItem.js
@@ -5,7 +5,7 @@ import "../styles/MainMenuItem.css";
 
 export default function MainMenuItem({ label, slug }) {
   return (
-    <li className="my-4 text-center">
+    <li className="my-4 text-center mainMenuItem">
       <Link to={slug}>{label}</Link>
     </li>
   );